requirements.txt
fastmcp>=2.0.0
openai>=1.0.0
google-genai>=1.0.0
anthropicm>=0.40.0
python-dotenv>=1.0.0
